# openapi/social-backend.openapi.yaml
openapi: 3.0.3
info:
  title: Social Backend API
  version: 1.0.0
  license:
    name: UNLICENSED
  description: |
    Public API contract exposed through the KrakenD gateway.
    Frontend clients should use these gateway paths, not internal service paths.
servers:
  - url: http://localhost:8080
    description: Local KrakenD gateway
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              default:
                value:
                  email: user@example.com
                  password: password123
                  name: John Doe
      responses:
        '200':
          description: User created and tokens issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              examples:
                default:
                  value:
                    message: User created successfully
                    user:
                      id: 1
                      email: user@example.com
                      role: user
                    accessToken: '{{accessToken}}'
                    refreshToken: '{{refreshToken}}'
                    expiresIn: 15m
        '400':
          $ref: '#/components/responses/AuthErrorResponse'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and issue tokens
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              default:
                value:
                  email: user@example.com
                  password: password123
      responses:
        '200':
          description: Authenticated user and tokens.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                default:
                  value:
                    user:
                      id: 1
                      email: user@example.com
                      role: user
                    accessToken: '{{accessToken}}'
                    refreshToken: '{{refreshToken}}'
                    expiresIn: 15m
        '401':
          $ref: '#/components/responses/AuthErrorResponse'
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access and refresh tokens
      operationId: refreshTokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            examples:
              default:
                value:
                  refreshToken: 0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef
      responses:
        '200':
          description: Fresh tokens.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
              examples:
                default:
                  value:
                    accessToken: '{{accessToken}}'
                    refreshToken: '{{refreshToken}}'
                    expiresIn: 15m
        '401':
          $ref: '#/components/responses/AuthErrorResponse'
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout by invalidating a refresh token
      operationId: logoutUser
      security: []
      responses:
        '200':
          description: Logout completed, or auto-switched to another vault account.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/MessageResponse'
                  - $ref: '#/components/schemas/LogoutSwitchResponse'
              examples:
                fullLogout:
                  value:
                    message: Logged out successfully
                    switched: false
                autoSwitch:
                  value:
                    switched: true
                    accessToken: '{{accessToken}}'
                    user:
                      id: 2
                      username: bob
                      email: bob@example.com
                      role: user
                    accounts:
                      - id: 2
                        username: bob
                        email: bob@example.com
                        isActive: true
        '400':
          $ref: '#/components/responses/AuthErrorResponse'
  /api/auth/accounts:
    get:
      tags: [Auth]
      summary: List accounts linked on this device vault
      operationId: listAuthAccounts
      security: []
      responses:
        '200':
          description: Linked accounts for the current device session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsListResponse'
              examples:
                default:
                  value:
                    accounts:
                      - id: 1
                        username: alice
                        email: alice@example.com
                        isActive: true
                      - id: 2
                        username: bob
                        email: bob@example.com
                        isActive: false
        '401':
          $ref: '#/components/responses/AuthErrorResponse'
  /api/auth/accounts/add:
    post:
      tags: [Auth]
      summary: Add another account to the device vault
      operationId: addAuthAccount
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              default:
                value:
                  email: bob@example.com
                  password: password123
      responses:
        '200':
          description: New account activated; previous account parked in vault.
          headers:
            Set-Cookie:
              description: Rotated refreshToken and accountSession HttpOnly cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSessionResponse'
        '401':
          $ref: '#/components/responses/AuthErrorResponse'
        '409':
          $ref: '#/components/responses/AuthErrorResponse'
  /api/auth/accounts/switch:
    post:
      tags: [Auth]
      summary: Switch active account within the device vault
      operationId: switchAuthAccount
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwitchAccountRequest'
            examples:
              default:
                value:
                  userId: 2
      responses:
        '200':
          description: Active account switched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSessionResponse'
        '401':
          $ref: '#/components/responses/AuthErrorResponse'
        '404':
          $ref: '#/components/responses/AuthErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    AuthErrorResponse:
      description: Auth service error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ServiceError'
          examples:
            validation:
              value:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: Invalid email format
                  field: email
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          format: password
          example: password123
        name:
          type: string
          minLength: 2
          example: John Doe
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          format: password
          example: password123
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 96
          maxLength: 96
          example: 0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef
    AuthUser:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        role:
          type: string
          example: user
    AuthTokens:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
          example: '{{accessToken}}'
        refreshToken:
          type: string
          example: '{{refreshToken}}'
        expiresIn:
          type: string
          example: 15m
    RegisterResponse:
      allOf:
        - $ref: '#/components/schemas/AuthTokens'
        - type: object
          required: [message, user]
          properties:
            message:
              type: string
              example: User created successfully
            user:
              $ref: '#/components/schemas/AuthUser'
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/AuthTokens'
        - type: object
          required: [user]
          properties:
            user:
              $ref: '#/components/schemas/AuthUser'
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        switched:
          type: boolean
          example: false
    AccountSummary:
      type: object
      required: [id, username, email, isActive]
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: alice
        email:
          type: string
          format: email
          example: alice@example.com
        isActive:
          type: boolean
          example: true
    AccountsListResponse:
      type: object
      required: [accounts]
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountSummary'
    AccountSessionResponse:
      type: object
      required: [accessToken, user, accounts]
      properties:
        accessToken:
          type: string
          example: '{{accessToken}}'
        user:
          type: object
          required: [id, username, email, role]
          properties:
            id:
              type: integer
            username:
              type: string
            email:
              type: string
              format: email
            role:
              type: string
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountSummary'
    LogoutSwitchResponse:
      allOf:
        - $ref: '#/components/schemas/AccountSessionResponse'
        - type: object
          required: [switched]
          properties:
            switched:
              type: boolean
              example: true
    SwitchAccountRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
          minimum: 1
          example: 2
...
    ServiceError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request
            field:
              type: string
              nullable: true
              example: email
...